pyproject.toml
[project]
name = "agent-forge"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "websockets>=13.0",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.3"]
slack = ["slack-bolt>=1.18", "slack-sdk>=3.27"]
media = ["openai-whisper>=20231117"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.27"]

[project.scripts]
forge = "agent_forge.cli:main"
agent-forge = "agent_forge.main:cli"

[tool.setuptools.packages.find]
include = ["agent_forge*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
